BB3D Occlusion type is now int

+++ paralleldomain/common/dgp/v0/dtos.py
@@ -204,7 +204,7 @@ class BoundingBox3DBoxDTO:
     width: float
     length: float
     height: float
-    occlusion: float
+    occlusion: int
     truncation: float
